financa/bot/requirements.txt
python-telegram-bot[webhooks]==21.3
httpx==0.27.0
python-dotenv==1.0.1
